feat: 클럽 조인 여부에 따라 조건부 렌더링
